+++ lib/src/widgets/selection.dart
@@ -5,8 +5,8 @@
 
 import 'dart:math' as math;
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:nt4f04unds_widgets/nt4f04unds_widgets.dart';
-import 'package:nt4f04unds_widgets/src/constants.dart';
 
 /// A regular [AppBar] which to use with selection and [SelectionController].
 /// 
@@ -33,10 +33,12 @@ class SelectionAppBar extends AppBar {
     double elevationSelection = 2.0,
     ShapeBorder? shape,
     Color? backgroundColor,
-    Brightness? brightness,
+    SystemUiOverlayStyle? systemOverlayStyle,
     IconThemeData? iconTheme,
     IconThemeData? actionsIconTheme,
     TextTheme? textTheme,
+    TextStyle? toolbarTextStyle,
+    TextStyle? titleTextStyle,
     bool primary = true,
     bool? centerTitle,
     bool excludeHeaderSemantics = false,
@@ -85,10 +87,11 @@ class SelectionAppBar extends AppBar {
         elevation: selectionController.inSelection ? elevationSelection : elevation,
         shape: shape,
         backgroundColor: backgroundColor,
-        brightness: brightness,
+        systemOverlayStyle: systemOverlayStyle,
         iconTheme: iconTheme,
         actionsIconTheme: actionsIconTheme,
-        textTheme: textTheme,
+        toolbarTextStyle: toolbarTextStyle,
+        titleTextStyle: titleTextStyle,
         primary: primary,
         centerTitle: centerTitle,
         excludeHeaderSemantics: excludeHeaderSemantics,